requirements.txt
google-genai>=1.19.0
pandas>=1.5
numpy>=1.26
tqdm>=4.66
PyYAML>=6.0
python-dateutil>=2.8
python-dotenv>=1.0.0
openai>=1.86.0
regex>=2023.10
scikit-learn>=1.3
ruamel.yaml>=0.17
spacy>=3.7,<3.9
spacy-langdetect==0.1.2
# Added for v1.0.0-rc1 architecture refactoring
pydantic-settings>=2.0.0
typer>=0.9.0
structlog>=23.0.0 